move more directories for better project structure
