+++ edn-react/src/features/chat/services/ChatService.ts
@@ -1,4 +1,4 @@
 import { signal } from "@preact/signals-react";
-import { ActivityItem } from "../features/chat/types/activityItem";
+import { ActivityItem } from "../types/activityItem";
 
 export const messages = signal([] as ActivityItem[]);
