+++ edn-react/src/pages/Chat.tsx
@@ -1,8 +1,8 @@
 import { GradientBGTop, GradientBGBottom } from "../components/gradient/index";
 import { ChatInputComponent, ChatOutputComponent } from "../features/chat";
-import { useAuthFetchData } from "../services/AuthFetch";
+import { useAuthFetchData } from "../features/authentication/services/AuthFetch";
 import { CreateCompletionsResponse } from "../interfaces/openAI/interfaces";
-import { messages } from "../services/ChatService";
+import { messages } from "../features/chat/services/ChatService";
 import { ActivityItem } from "../features/chat/types/activityItem";
 
 function createMessage(response: CreateCompletionsResponse | null) {
